pyproject.toml
[project]
dependencies = [
  "matplotlib>=3.10.8",
  "numpy>=2.4.2",
  "pandas>=3.0.1",
  "psycopg2-binary>=2.9.11",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5",
]
description = "Download, parse, and store PostgreSQL mailing list mbox archives into Postgres"
name = "pginbox"
requires-python = ">=3.12"
version = "0.1.0"
